requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
PyMuPDF==1.23.23
google-generativeai
